requirements.txt
pytelegrambotapi
psycopg2-binary
openai
python-dotenv
fastapi
uvicorn
